requirements.txt
torch>=2.0.0
transformers>=4.30.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pytest>=7.4.0
python-dotenv>=1.0.0

# Web API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Whisper
openai-whisper>=20231117

# Auth & database
sqlalchemy>=2.0.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0